flaskService/requirements.txt
Flask                   ==  1.0.2
Flask-RESTful           ==  0.3.7
flask-jwt-extended      ==  3.21.0
Flask-SQLAlchemy        ==  2.4.1
marshmallow             ==  3.2.1
flask-marshmallow       ==  0.10.1
marshmallow-sqlalchemy  ==  0.19.0
bcrypt                  ==  3.1.7 
#mysql-connector         ==  2.2.9
Werkzeug                ==  0.15.5
gunicorn                == 19.9.0
#flask-sqlacodegen
#sqlacodegen
pymysql
